+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased 
+
+* Remove onPauseStarted and onPauseEnded signals from AsyncLineView (they were not functioning anyway in prior 0.3 releases)
+
 ## [0.3.0-beta 4] 2025-02-19
 
 * Add a way to set Saliency Strategy from godot without having to write C# with @misha-cilantro (#85)